renaming splitOn() to groupBy()
on the assumption that java devs are more familiar with databases than with statistical software, and that the naming in statitistical software is inconsistent in any case.

OTOH, this is not exactly like SQL group by, which is bound up with summarization and does not return the intermediate groups as this does.