Made ColumnType an interface (#315)
* Made ColumnType an interface, so that it may be easier in the future to
create new column types, and perhaps may be possible for library users to add new column types specific to their needs.
Currently, this cannot be done because there are several key statements that switch on column type. It is a future enhancement to
eliminate those switch statements by (if possible) moving the requisite logic to the column types.  This would be more
object-oriented anyway

* codacy concern addressed

* replace list-based ColumnType registry with Map version for easier name-based access

* moved the implementation class for column types to the column package (and out of the API package)

* created specific classes for each type

* made the map from sql types to column types mutable and added public method to update