Improve interpretation of NUMBER SQL types.
When a database ResultSet is used to instantiate a Table, the current
interpretation of the NUMBER type was to automatically map it to a
DoubleColumn, but Oracle converts all of its convenient types like
INTEGER to a NUMBER type internally, and that NUMBER type is the type
seen by tablesaw during import of the ResultSet.  This change interprets
the NUMBER type according to the precision and scale properties to map
it to the appropriate column type.

 On branch sql-number-type-mapping
 Changes to be committed:
	modified:   core/src/main/java/tech/tablesaw/io/jdbc/SqlResultSetReader.java
	modified:   core/src/test/java/tech/tablesaw/columns/ColumnTest.java
	modified:   core/src/test/java/tech/tablesaw/io/jdbc/SqlResultSetReaderTest.java
	modified:   core/src/test/java/tech/tablesaw/util/TestDb.java