Fix for issue #443
byte[] was not initialized if an input stream was used for reading.

Added tests that use readers and streams on a local file so they can be run in the regression suite.

Added a check to ensure that CsvReadOptions Builder has only one data source configured (stream, reader, or file).