partial implementation of "Better handling of file encodings" #638
Detects encodings for input text files when possible, and when no encoding is specifically provided as an argument. If the detection confidence is weak, it falls back on the system default charset.

This implementation adds a new dependency to Core, which is undesireable, but getting data into tablesaw correctly is arguably it's most important function so I think it's worth it.