fix bug in appendCell that caused custom parser to be ignored
In TextColumn, calling appendCell(string, parser) failed because the results of the call to parser.parse(string) was passed to appendObj(), which called parse again, this time using the default parser. This is fixed.

This commit also makes it possible to directly set the missingValueStrings list in a custom parser.