first commit
This branch is for evaluating the possibility of using ColumnPrintFormatters to optionally format output.

There will be a writeOption called ```usePrintFormatters(boolean)``` that turns the logic on or off. It will be off by default, matching the current status.

The print formatter is set on the column and would then be used by both table.print and table.write, providing more consistent output

PrintFormatters have the advantage of handling missing values so the user does not have to worry about them blowing up a Java NumberFormatter for instance, while allowing the user to set the missing value symbol for each column.