Update moneyball tutorial (#1052)
* Add Gradle smile dependency

* Add link to baseball.csv

The only reference to where the baseball.csv file is located is the online course. You have to sign up, etc in order to access that location. Seeing as the same file is in the tablesaw repo already, I provided a link in the tutorial for easy download.

* Update code snippet syntax

Update to new tablesaw syntax.

* Correct spacing

Some tabs were inserted during copy/paste

* URL " " to %20

Replace all blank space in url with safety character %20 to ensure images load properly.

* Add link to tablesaw-core and tablesaw-jsplot 

Added links to tablesaw dependencies to make this a bit more beginner-friendly.

* Small formatting changes